gnec_hackathon/backend/asl-service/requirements.txt
fastapi==0.104.0
uvicorn==0.24.0
python-multipart==0.0.6
numpy==1.24.3
opencv-python==4.8.0.76
torch==2.0.1
mediapipe>=0.10.5
websockets==11.0.3
python-dotenv==1.0.0
requests==2.31.0
google-generativeai>=0.3.0 